requirements.txt
# Data & API
refinitiv-data>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Configuration
pyyaml>=6.0

# Analysis & Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
